requirements.txt
decord==0.6.0
pillow==11.2.1
